Registration Bugs Fixed + transaction date bug fixed
here i have solved the bug when the user logout, the application crash. Fixed.
and now when the user logout all the unecessary controllers will be deleted too, for the performance.
the date in the transaction page was showing the current date everytime u leave the page and return, while the data shown are not for that date. Fixed by using the same date from the transaction controller.